Changed theme of the game
Changed the main character to be an owl. Additionally, I changed the name of the game from "petmouse" to "petowl". I increased the scale of the owl on click, as well as the rotation of its movements. I also increased the size of the droppings, and the playable area. Finally I changed the background color from white to cyan.